Separate ConsolePal.GetCharset into a reusable Common helper
To allow it to be used from other components.
